refactor: move guardrails helpers to shared
