--- extensions/guardrails/utils/warnings.ts
@@ -1,5 +1,0 @@
-export {
-  addPendingWarning,
-  drainPendingWarnings,
-  getPendingWarnings,
-} from "../../../src/shared/warnings";
